@@ -9,24 +9,23 @@
     "build": "webpack --progress"
   },
   "devDependencies": {
-    "babel-loader": "^6.4.1",
-    "babel-preset-es2015": "^6.24.0",
-    "babel-preset-react": "^6.23.0",
-    "halogen": "^0.2.0",
-    "webpack": "^2.2.1",
-    "webpack-dev-server": "^2.4.2"
+    "babel-core": "^6.24.1",
+    "babel-loader": "^7.0.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "html-webpack-plugin": "^2.28.0",
+    "webpack": "^2.5.1",
+    "webpack-dev-server": "^2.4.5"
   },
   "dependencies": {
-    "babel-core": "^6.24.0",
-    "html-webpack-plugin": "^2.28.0",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "react-loadable": "^3.0.1",
-    "react-router": "^4.0.0",
-    "react-router-dom": "^4.0.0",
-    "react-spinner": "^0.2.6",
-    "styled-components": "^1.4.4",
-    "webpack-dev-server": "^2.4.2",
-    "react-listener-provider": "^0.1.2"
+    "prop-types": "^15.5.8",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-listener-provider": "^0.2.0",
+    "react-loadable": "^3.3.1",
+    "react-router": "^4.1.1",
+    "react-router-dom": "^4.1.1",
+    "react-spinner": "^0.2.7",
+    "styled-components": "^1.4.6"
   }
 }
