@@ -1,6 +1,5 @@
 import React from 'react';
 import Loadable from 'react-loadable';
-import Loader from 'halogen/PulseLoader';
 
 function fakeDelay(ms) {
   return new Promise(resolve => {
@@ -10,5 +9,5 @@ function fakeDelay(ms) {
 
 export default Loadable({
   loader: () => fakeDelay(2000).then(() => System.import('./Home')),
-  LoadingComponent: () => <Loader color="#333333" size="16px" />
-})
+  LoadingComponent: () => <div>Loading...</div>
+});
